Cake\I18n\Time toQuarter() change location Int cast
GitHub Issue #12903

+++ src/I18n/Time.php
@@ -172,9 +172,9 @@ class Time extends MutableDateTime implements JsonSerializable
      */
     public function toQuarter($range = false)
     {
-        $quarter = ceil($this->format('m') / 3);
+        $quarter = (int)ceil($this->format('m') / 3);
         if ($range === false) {
-            return (int)$quarter;
+            return $quarter;
         }
 
         $year = $this->format('Y');
